Move neighbor code into own file
